requirements.txt
fastapi
uvicorn
PyPDF2
requests
python-dotenv
langchain-text-splitters
together
pinecone-client
